Splitting Components by Feature and State - more lean App.jsx
