+++ src/App.jsx
@@ -1,53 +1,16 @@
-import { useState } from 'react';
-
-import { CORE_CONCEPTS } from './data.js';
 import Header from './components/Header/Header.jsx';
-import CoreConcept from './components/CoreConcept.jsx';
-import TabButton from './components/TabButton.jsx';
-import { EXAMPLES } from './data.js';
+import CoreConcepts from './components/CoreConcepts.jsx';
+import Examples from './components/Examples.jsx';
+
 
 function App() {
-
-  const [selectedTopic, setSelectedTopic] = useState();
-
-  function handleSelect(selectedButton) {
-    //selectedButton => 'components','jsx','props','state'
-    setSelectedTopic(selectedButton);
-    //console.log(selectedTopic);
-  }
-
-  //console.log('APP COMPONENT EXECUTING');
-
-  let tabContent = <p>Please select a topic</p>;
-
-  if (selectedTopic) {
-    tabContent =
-      <div id="tab-content">
-        <h3>{EXAMPLES[selectedTopic].title}</h3>
-        <p>{EXAMPLES[selectedTopic].description}</p>
-        <pre>
-          <code>
-            {EXAMPLES[selectedTopic].code}
-          </code>
-        </pre>
-      </div>
-  }
 
   return (
     <>
       <Header />
       <main>
+        <CoreConcepts />
         
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected={selectedTopic === 'components'} onSelect={() => handleSelect('components')}>Components</TabButton>
-            <TabButton isSelected={selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
-            <TabButton isSelected={selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
-            <TabButton isSelected={selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
-          </menu>
-          {tabContent}
-        </section>
       </main>
     </>
   );
